fix(web): break client.ts ↔ conversation-sse circular import
CI's Unit Tests job (verify:test-unit) crashed in the web suite with
'Export named getActiveWorkspaceId/setActiveWorkspaceId/ApiClientError not
found in module web/src/api/client.ts' — even though those exports exist.

Root cause: a circular import. client.ts imported getConversationSubscriberId
from conversation-sse.ts, and conversation-sse.ts imported refreshSession back
from client.ts. Bun resolves cycles in module-evaluation order, which is
non-deterministic under the parallel test runner; when a bridge test triggered
conversation-sse first, client.ts's named exports weren't initialized yet and
linking failed. Latent on main; deterministically tripped on this branch once
the module graph shifted.

Fix: extract the per-conversation subscriber-id registry (a Map + get/set/clear,
no other deps) into web/src/api/conversation-subscribers.ts. client.ts and
conversation-sse.ts now both depend only on that leaf module — no cycle. Pure
move, no behavior change.
