+++ web/src/api/conversation-sse.ts
@@ -10,35 +10,10 @@
  */
 
 import { refreshSession } from "./client";
-
-/**
- * Per-conversation `subscriberId` registry.
- *
- * The server-issued subscriber id arrives in the first SSE frame
- * (`event: subscribed`) once a conversation event stream opens. We
- * stash it here so the chat-stream POST path can pick it up via
- * `getConversationSubscriberId(convId)` and forward it as
- * `X-Origin-Subscriber-Id` — that makes the broadcast skip this
- * tab's own conv-events subscription and prevents the sender from
- * double-handling every event (once via the chat-stream HTTP
- * response, once via the broadcast hitting its own subscription).
- *
- * Cleared on stream cancel / close to avoid stale ids leaking into
- * a future stream attempt for the same conv.
- */
-const conversationSubscriberIds = new Map<string, string>();
-
-export function getConversationSubscriberId(conversationId: string): string | undefined {
-  return conversationSubscriberIds.get(conversationId);
-}
-
-function setConversationSubscriberId(conversationId: string, subscriberId: string): void {
-  conversationSubscriberIds.set(conversationId, subscriberId);
-}
-
-function clearConversationSubscriberId(conversationId: string): void {
-  conversationSubscriberIds.delete(conversationId);
-}
+import {
+  clearConversationSubscriberId,
+  setConversationSubscriberId,
+} from "./conversation-subscribers";
 
 /** Options for connecting to a conversation event stream. */
 export interface ConversationSseOptions {
